test: add integration tests and some refactoring
