+++ test/mocks/external.ts
@@ -1,0 +1,5 @@
+export const windowMock: Window = {
+  navigator: {
+    userAgent: 'TestAgent'
+  }
+} as Window
