+++ test/specs/LeanplumInternal.test.ts
@@ -16,46 +16,11 @@
 
 import Constants from '../../src/Constants'
 import LeanplumInternal from '../../src/LeanplumInternal'
-import LeanplumRequest from '../../src/LeanplumRequest'
-import VarCache from '../../src/VarCache'
-import { startResponse } from '../data/responses'
+import { APP_ID, KEY_DEV } from '../data/constants'
+import { windowMock } from '../mocks/external'
+import { lpRequestMock, varCacheMock } from '../mocks/internal'
 
-// Test data
-const APP_ID = 'app_BWTRIgOs0OoevDfSsBtabRiGffu5wOFU3mkxIxA7NBs'
-const KEY_DEV = 'dev_Bx8i3Bbz1OJBTBAu63NIifr3UwWqUBU5OhHtywo58RY'
-const KEY_PROD = 'prod_A1c7DfHO6XTo2BRwzhkkXKFJ6oaPtoMnRA9xpPSlx74'
-
-const windowMock: Window = {
-  navigator: {
-    userAgent: 'TestAgent'
-  }
-} as Window
-
-const lpRequestMock: Partial<jest.Mocked<LeanplumRequest>> = {
-  getLastResponse: jest.fn(),
-  isResponseSuccess: jest.fn().mockImplementation((response) => Boolean(response?.success)),
-  request: jest.fn()
-}
-
-const varCacheMock: Partial<jest.Mocked<VarCache>> = {
-  applyDiffs: jest.fn(),
-  loadDiffs: jest.fn(),
-  getVariable: jest.fn(),
-  getVariables: jest.fn(),
-  setVariables: jest.fn(),
-  getVariantDebugInfo: jest.fn(),
-  setVariantDebugInfo: jest.fn(),
-  addVariablesChangedHandler: jest.fn(),
-  removeVariablesChangedHandler: jest.fn(),
-  clearUserContent: jest.fn(),
-}
-
-jest.mock('../../src/LeanplumRequest', () => {
-  return jest.fn().mockImplementation(() => {
-    return lpRequestMock
-  })
-})
-
+jest.mock('../../src/LeanplumRequest', () => jest.fn().mockImplementation(() => lpRequestMock))
 jest.mock('../../src/VarCache', () => jest.fn().mockImplementation(() => varCacheMock))
 
 describe(LeanplumInternal, () => {
@@ -281,37 +246,6 @@ describe(LeanplumInternal, () => {
 
         expect(varCacheMock.removeVariablesChangedHandler).toHaveBeenCalledTimes(1)
         expect(varCacheMock.removeVariablesChangedHandler).toHaveBeenCalledWith(handler)
-      })
-    })
-
-    describe('Flows', () => {
-      it('calls variable changed callback after start', (done) => {
-        const userId = '123'
-        const variables = {
-          gender: 'female',
-          age: 27
-        }
-
-        lpRequestMock.getLastResponse.mockImplementation(() => {
-          return startResponse.response[0]
-        })
-        lpRequestMock.request.mockImplementation((action, args, opts) => {
-          if (action === Constants.METHODS.START) {
-            if (opts?.response) {
-              opts.response(startResponse)
-            }
-          }
-        })
-
-        lp.setVariables(variables)
-        lp.addVariablesChangedHandler(() => {
-          expect(lp.getVariables().gender).toBe(variables.gender)
-          expect(lp.getVariables().age).toBe(variables.age)
-        })
-        lp.start(userId, {}, (success) => {
-          expect(success).toBe(true)
-          return done(success ? null : success)
-        })
       })
     })
   })
